Seperating growth and division from the main code.
